refactor: isolate frontend bootstrap decisions
